# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.5"
  title: SiB Colombia - Catalogue API
  description: API for SiB Colombia Catalogue. This API is used as a backend for consult the information of the species records reported in the SiB Colombia.
  termsOfService: http://swagger.io/terms/
  contact:
    name: SiB Colombia Developer Team
    url: http://www.sibcolombia.net
    email: oduque@humboldt.org.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /api/v1.5
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /records/{id}/abstract/{version}:
    x-swagger-router-controller: abstract
    get: 
      description: Returns a version of a Abstract element by record id and version
      tags:
        - element
      operationId: getAbstract
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: version
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AbstractVersion"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/{id}/abstract:
    x-swagger-router-controller: abstract
    post:
      description: Save a new version of a abstract related with the a recrd with the id
      tags:
        - element
      operationId: postAbstract
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: abstractVersion
          description: abstractVersion object
          in: body
          required: true
          schema:
            $ref: '#/definitions/AbstractVersion'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResultResponse"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
  /records/to_review/{id}/abstract:
    x-swagger-router-controller: abstract
    get: 
      description: Returns a list of Abstract version with state "to_review"
      tags:
        - element
      operationId: getToReviewAbstract
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AbstractVersionList"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/accepted/{id}/abstract:
    x-swagger-router-controller: abstract
    get: 
      description: Returns the abstract version accepted #**
      tags:
        - element
      operationId: getLastAcceptedAbstract
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/AbstractVersion"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/accepted/{id}/abstract/{version}:
    x-swagger-router-controller: abstract
    put: 
      description: Update a AbstractVersion state to accepted
      tags:
        - element
      operationId: setAcceptedAbstract
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: version
          type: integer
          in: path
          required: true
        - name: user
          description: User object
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        "200":
          description: Success
          
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  AbstractVersionList:
    properties:
      abstractList:
        type: array
        items:
          $ref: '#/definitions/AbstractVersion'
...
  AbstractVersion:
    required:
      - abstract
      - id_user
    properties:
      _id:
        type: string
      id_record: 
        type: string
      created: 
        type: string #****
      id_user:
        type: string
      version:
        type: integer
      state:
        type: string
      element:
        type: string
      abstract:
        type: string
...
  User:
    required:
      - id_user
    properties:
      id_user:
        type: string
        pattern: '^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$'
  ResultResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string